+++ actix-limitation/src/errors.rs
@@ -1,25 +1,25 @@
-use derive_more::{Display, Error, From};
+use derive_more::{Display, Error as DeriveError, From};
 
 use crate::status::Status;
 
 /// Failure modes of the rate limiter.
-#[derive(Debug, Display, Error, From)]
+#[derive(Debug, Display, DeriveError, From)]
 pub enum Error {
     /// Redis client failed to connect or run a query.
-    #[display(fmt = "Redis client failed to connect or run a query")]
+    #[display("Redis client failed to connect or run a query")]
     Client(redis::RedisError),
 
     /// Limit is exceeded for a key.
-    #[display(fmt = "Limit is exceeded for a key")]
+    #[display("Limit is exceeded for a key")]
     #[from(ignore)]
     LimitExceeded(#[error(not(source))] Status),
 
     /// Time conversion failed.
-    #[display(fmt = "Time conversion failed")]
+    #[display("Time conversion failed")]
     Time(time::error::ComponentRange),
 
     /// Generic error.
-    #[display(fmt = "Generic error")]
+    #[display("Generic error")]
     #[from(ignore)]
     Other(#[error(not(source))] String),
 }