+++ actix-session/src/storage/session_key.rs
@@ -45,7 +45,7 @@ impl From<SessionKey> for String {
 }
 
 #[derive(Debug, Display, From)]
-#[display(fmt = "The provided string is not a valid session key")]
+#[display("The provided string is not a valid session key")]
 pub struct InvalidSessionKeyError(anyhow::Error);
 
 impl std::error::Error for InvalidSessionKeyError {