+++ actix-settings/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update `derive_more` dependency to `1.0`.
+
 ## 0.8.0
 
 - Add `openssl` crate feature for TLS settings using OpenSSL.