+++ actix-settings/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Update `toml` dependency to `0.9`.
 - Minimum supported Rust version (MSRV) is now 1.80.
 
 ## 0.8.0