+++ actix-settings/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.9.0
+
 - Update `toml` dependency to `0.9`.
 - Minimum supported Rust version (MSRV) is now 1.80.